+++ src/index.ts
@@ -26,7 +26,7 @@ export default class GameGuardClient {
   * 
   * @property {WebSocket}
   */
- private _socket!: WebSocket;
+  private _socket!: WebSocket;
 
   /**
    * A reference to the ClientData module.
@@ -71,6 +71,15 @@ export default class GameGuardClient {
   private _disconnected: Hypergiant = new Hypergiant();
 
   /**
+   * This client's latency to the server, in milliseconds.
+   * 
+   * @private
+   * 
+   * @property {number}
+   */
+  private _latency: number = 0;
+
+  /**
    * @param {Object} [options] The initialization parameters passed to this instance.
    * @param {boolean} [options.secure=false] Indicates whether the websocket will connect to the server with a secure connection or not.
    */
@@ -100,6 +109,13 @@ export default class GameGuardClient {
    * @returns {Hypergiant}
    */
   get disconnected(): Hypergiant { return this._disconnected; }
+
+  /**
+   * Returns this client's latency to the server, in milliseconds.
+   * 
+   * @returns {number}
+   */
+  get latency(): number { return this._latency; }
 
   /**
    * Initialize the WebSocket connection and all of the events that we need to respond to.
@@ -146,7 +162,17 @@ export default class GameGuardClient {
 
     const msg: Message = new Message(parsed.type, parsed.contents);
 
-    this.messaged.dispatch(msg);
+    switch (msg.type) {
+      case 'latency-ping':
+        const latencyMessage = new Message('latency-pong', msg.contents);
+
+        this._socket.send(latencyMessage.stringify());
+        break;
+      case 'latency':
+        this._latency = parseFloat(msg.contents);
+      default:
+        this.messaged.dispatch(msg);
+    }
   }
 
   /**